backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
deprecation==2.1.0
distro==1.9.0
fastapi==0.128.0
fsspec==2026.1.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langsmith==0.6.6
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==2.4.1
openai==2.16.0
orjson==3.11.5
packaging==25.0
pillow==12.1.0
portalocker==3.2.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.3.2
pypdf==6.6.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
PyYAML==6.0.3
qdrant-client==1.16.2
realtime==2.27.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zstandard==0.25.0
